HCIToolbox

HCIToolbox.Mean

HCIToolbox.Median

HCIToolbox.NMF

HCIToolbox.PCA

HCIToolbox.Pairet

Base.reduce

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.design

HCIToolbox.mask

HCIToolbox.mask!

Statistics.mean

Statistics.median

HCIToolbox.Mean — Type

Mean

Design using the mean of the cube

Examples

julia> cube = DataCube(ones(100, 100, 30));

julia> design(Mean, cube)
(A = [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0; … ; 1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0], w = UniformScaling{Bool}(true), S = [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0; … ; 1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0])

See Also

Median

HCIToolbox.Median — Type

Median

Design using the median of the cube

Examples

julia> cube = DataCube(ones(100, 100, 30));

julia> design(Median, cube)
(A = [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0; … ; 1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0], w = UniformScaling{Bool}(true), S = [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0; … ; 1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0])

See Also

Mean

HCIToolbox.NMF — Type

NMF

HCIToolbox.PCA — Type

PCA

Use principal component analysis (PCA) to reduce data cube. ncomponents defines how many principal components to use.

Uses MultivariateStats.PCA for decomposition. See MultivariateStats.fit(PCA; ...) for keyword arguments

Arguments

ncomps::Int = nframes(::DataCube) - The number of components to keep. Cannot be larger than the number of frames in the DataCube.=

Examples

julia> cube = DataCube(ones(100, 100, 30));

julia> design(PCA, cube)
(A = [1.0; 0.0; … ; 0.0; 0.0], w = [0.0 0.0 … 0.0 0.0], S = [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0; … ; 1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0])

HCIToolbox.Pairet — Type

Pairet{<:Union{PCA, NMF}}

Base.reduce — Method

reduce(::Type{<:Design}, ::DataCube, args...; collapse=median, kwargs...)

Using a given Design, will reduce the DataCube by first finding the approximate reconstruction with design and then derotating and collapsing (using whichever function specified by collapse). Any kwargs will be passed to design.

HCIToolbox.derotate! — Method

derotate!(::DataCube)

In-place version of derotate

HCIToolbox.derotate — Method

derotate(::DataCube)

De-rotates a DataCube using its internal angles.

HCIToolbox.design — Method

design(::Type{<:Design}, ::DataCube, args...; kwargs...)

Create a design matrix and weights from the given DataCube. The kwargs will vary based on the design algorithm.

Returns

The output of a design matrix will be a named tuple with 3 parameters:

A - The design Matrix

w - The weight vector (the transform of our data cube)

S - The reconstruction of our data cube (usually A * w)

HCIToolbox.mask! — Method

mask!(::AbstractMatrix, npix)

In-place version of mask

HCIToolbox.mask — Method

mask(::AbstractMatrix, npix)

Mask the inner-circle of an image with radius npix.

Statistics.mean — Method

Statistics.mean(::DataCube)

Collapses a DataCube by finding the mean frame. Note: does not derotate.

Statistics.median — Method

Statistics.median(::DataCube)

Collapses a DataCube by finding the median frame. Note: does not derotate.