Spectral Processing

The following functions are used specifically for processing spectral differential imaging (SDI) tensors/cubes.

Index

HCIToolbox.invscale

HCIToolbox.invscale_and_collapse

HCIToolbox.scale

HCIToolbox.scale_and_stack

HCIToolbox.scale_list

API/Reference

HCIToolbox.invscale — Function

invscale(cube::AbstractArray{T,3}, scales[, size])

Linearly contract each frame in cube by the corresponding scale in scales. Uses bilinear interpolaiton internally. The output size can be specified or else we will choose the smallest size that contains the largest stretch.

invscale(frame::AbstractMatrix, scale[, size])

Linearly contract frame with the ratio scale. Uses bilinear interpolation internally. If size is specified, will return a crop view.

HCIToolbox.invscale_and_collapse — Function

invscale_and_collapse(stack_cube::AbstractArray{T,3}, scales, size; kwargs...)

Given an SDI tensor that has been stacked into a cube, invscales each spectral slice and combines with collapse. The output will be cropped to size.

HCIToolbox.scale_list — Function

scale_list(wavelengths)

Returns a list of scaling factors for aligning SDI tensors from a list of wavelengths.

HCIToolbox.scale — Function

scale(cube::AbstractArray{T,3}, scales[, size])

Linearly stretch each frame in cube by the corresponding scale in scales. Uses bilinear interpolaiton internally. The output size can be specified or else we will choose the smallest size that contains the largest stretch.

scale(frame::AbstractMatrix, scale[, size])

Linearly stretch frame with the ratio scale, padding symmetrically with zeros to match out_size. Uses bilinear interpolation internally. If size is specified, will return a padded view.

HCIToolbox.scale_and_stack — Function

scale_and_stack(spcube::AbstractArray{T,4}, scales)

Given a 4-D spectral ADI (SDI) tensor, scales each, temporal slice according to scales and then concatenates into a cube with nλ * nf frames.