Utilities

The following routines are general utility or otherwise don't fit into a category

Index

HCIToolbox.mask_annulus

HCIToolbox.mask_annulus!

HCIToolbox.mask_circle

HCIToolbox.mask_circle!

HCIToolbox.normalize_par_angles

HCIToolbox.normalize_par_angles!

API/Reference

HCIToolbox.mask_circle — Function

mask_circle(::AbstractMatrix, npix; fill=0)
mask_circle(::AbstractArray, npix; fill=0)

Mask the inner-circle of an image with radius npix with value fill. Note that the input type must be compatible with the fill value's type. If the input is a cube it will mask each frame individually.

See Also

mask_circle!

HCIToolbox.mask_circle! — Function

mask_circle!(::AbstractMatrix, npix; fill=0)
mask_circle!(::AbstractArray, npix; fill=0)

In-place version of mask_circle

HCIToolbox.mask_annulus — Function

mask_annulus(::AbstractMatrix, npix_in, npix_out; fill=0)

Mask an annular region of an image with inner-radius npix_in, outer-radius npix_out with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_annulus!

HCIToolbox.mask_annulus! — Function

mask_annulus!(::AbstractMatrix, npix_in, npix_out; fill=NaN)

In-place version of mask_annulus

HCIToolbox.normalize_par_angles — Function

normalize_par_angles(angles)

Ensures parallactic angle list (in degrees) is positive monotonic with no jumps greater than 180°.

Examples

julia> normalize_par_angles([-10, 20, 190])
3-element Array{Int64,1}:
 350
  20
 190

HCIToolbox.normalize_par_angles! — Function

normalize_par_angles!(angles)

In-place version of normalize_par_angles