API/Reference

For now, here is a dump of all documented functions and types.

Index

HCIToolbox.combine

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.expand

HCIToolbox.flatten

HCIToolbox.mask_circle

HCIToolbox.mask_circle!

HCIToolbox.rotate

HCIToolbox.rotate!

API/Reference

HCIToolbox.combine — Method

combine(cube, [angles]; method=median)

Combine all the frames of a cube using method. If angles are provided, will use derotate! before combining.

Note that with angles this is slightly slower than doing combine(derotate!(cube, angles)), if you can overwrite the input.

Examples

julia> X = ones(2, 3, 3)
2×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 2] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 3] =
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X, [0, 90])
1×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0

[:, :, 2] =
 NaN  1.0  NaN

[:, :, 3] =
 1.0  1.0  1.0

HCIToolbox.derotate! — Method

derotate!(cube, angles)

In-place version of derotate

HCIToolbox.derotate — Method

derotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i counter-clockwise by the amount deg2rad(angles[i]).

See Also

derotate!, rotate, rotate!

HCIToolbox.expand — Method

expand(matrix)

Given a matrix of size (n, z), returns a cube of size (n, x, x) where x=√z.

Will throw an error if z is not a perfect square.

Examples

julia> X = ones(3, 4)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> expand(X)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

See Also

flatten

HCIToolbox.flatten — Method

flatten(cube)

Given a cube of size (n, x, y) returns a matrix with size (n, x * y) where each row is a flattened image from the cube.

Examples

julia> X = ones(3, 2, 2)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> flatten(X)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

See Also

expand

HCIToolbox.mask_circle! — Method

mask_circle!(::AbstractMatrix, npix; fill=NaN)

In-place version of mask_circle

HCIToolbox.mask_circle — Method

mask_circle(::AbstractMatrix, npix; fill=NaN)

Mask the inner-circle of an image with radius npix with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_circle!

HCIToolbox.rotate! — Method

rotate!(cube, angles)

In-place version of derotate

HCIToolbox.rotate — Method

rotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i clockwise by the amount deg2rad(angles[i]).

See Also

rotate!, derotate, derotate!