API/Reference

For now, here is a dump of all documented functions and types.

Index

HCIToolbox.collapse

HCIToolbox.collapse!

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.expand

HCIToolbox.flatten

HCIToolbox.get_annulus_segments

HCIToolbox.inject_image

HCIToolbox.inject_image

HCIToolbox.inject_image!

HCIToolbox.inject_image!

HCIToolbox.mask_annulus

HCIToolbox.mask_annulus!

HCIToolbox.mask_circle

HCIToolbox.mask_circle!

HCIToolbox.normalize_par_angles

HCIToolbox.normalize_par_angles!

HCIToolbox.shift_frame

HCIToolbox.shift_frame

HCIToolbox.shift_frame!

API/Reference

HCIToolbox.collapse! — Method

collapse!(cube, angles; method=median, deweight=true)

An in-place version of the derotating collapse. The only difference is in this version the cube will be derotated in-place.

HCIToolbox.collapse — Method

collapse(cube; method=median, fill=0)
collapse(cube, angles; method=median, deweight=true, fill=0)

Combine all the frames of a cube using method. If angles are provided, will use derotate before combining.

If deweight is true, the method of Bottom et al. 2017 will be used in which the combined image will be the derotated weighted sum of the frames weighted by the temporal variance. fill will be passed to derotate.

References

Bottom et al. 2017 "Noise-weighted Angular Differential Imaging"

Examples

julia> X = ones(2, 3, 3);

julia> collapse(X)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> collapse(X, [0, 90])
3×3 Array{Float64,2}:
 0.5  1.0  0.5
 1.0  1.0  1.0
 0.5  1.0  0.5

julia> collapse(X, [0, 90], fill=NaN)
3×3 Array{Float64,2}:
 NaN    1.0  NaN
   1.0  1.0    1.0
 NaN    1.0  NaN

See Also

collapse!

HCIToolbox.derotate! — Method

derotate!(cube, angles; fill=0)

In-place version of derotate which modifies cube.

HCIToolbox.derotate — Method

derotate(cube, angles; fill=0)

Rotates an array using the given angles in degrees.

This will rotate frame i counter-clockwise. Any values outside the original axes will be replaced with fill. If the given angles are true parallactic angles, the resultant cube will have each frame aligned with top pointing North.

Examples

julia> X = zeros(1, 3, 3); X[1, 1, 2] = 1;

julia> X[1, :, :]
3×3 Array{Float64,2}:
 0.0  1.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

julia> derotate(X, [90])[1, :, :]
3×3 Array{Float64,2}:
 0.0  3.22941e-16  0.0
 0.0  0.0          0.999991
 0.0  0.0          0.0

See Also

derotate!

HCIToolbox.expand — Method

expand(matrix)

Given a matrix of size (n, z), returns a cube of size (n, x, x) where x=√z.

Will throw an error if z is not a perfect square.

Examples

julia> X = ones(3, 4)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> expand(X)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

See Also

flatten

HCIToolbox.flatten — Method

flatten(cube)

Given a cube of size (n, x, y) returns a matrix with size (n, x * y) where each row is a flattened image from the cube.

Examples

julia> X = ones(3, 2, 2);

julia> flatten(X)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

See Also

expand

HCIToolbox.inject_image! — Method

inject_image!(frame, img; x, y, A=1)
inject_image!(frame, img; r, theta, A=1)

In-place version of inject_image which modifies frame.

HCIToolbox.inject_image! — Method

inject_image!(cube, img, [angles]; x, y, A=1)
inject_image!(cube, img, [angles]; r, theta, A=1)

In-place version of inject_image which modifies cube.

HCIToolbox.inject_image — Method

inject_image(frame, img; x, y, A=1)
inject_image(frame, img; r, theta, A=1)

Injects A * img into frame at the position relative to the center of frame given by the keyword arguments. If necessary, img will be bilinearly interpolated onto the new indices. When used for injecting a PSF, it is imperative the PSF is already centered and, preferrably, odd-sized.

Coordinate System

The positions are decided in this way:

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta - Parsed as polar coordinates from the center of the image. theta is a position angle.

Note

Due to the integral nature of array indices, frames or images with even-sized axes will have their center rounded to the nearest integer. This may cause unexpected results for small frames and images.

Examples

julia> inject_image(zeros(5, 5), ones(1, 1), A=2, x=2, y=1) # image coordinates
5×5 Array{Float64,2}:
 0.0  2.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> inject_image(zeros(5, 5), ones(3, 3), r=1.5, theta=90) # polar coords
5×5 Array{Float64,2}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0
 0.0  0.0  1.0  1.0  0.0

HCIToolbox.inject_image — Method

inject_image(cube, img, [angles]; x, y, A=1)
inject_image(cube, img, [angles]; r, theta, A=1)

Injects A * img into each frame of cube at the position given by the keyword arguments. If angles are provided, the position in the keyword arguments will correspond to the img position on the first frame of the cube, with each subsequent repositioned img being rotated by -angles in degrees. This is useful for fake companion injection.

HCIToolbox.mask_annulus! — Method

mask_annulus!(::AbstractMatrix, npix_in, npix_out; fill=NaN)

In-place version of mask_annulus

HCIToolbox.mask_annulus — Method

mask_annulus(::AbstractMatrix, npix_in, npix_out; fill=0)

Mask an annular region of an image with inner-radius npix_in, outer-radius npix_out with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_annulus!

HCIToolbox.mask_circle! — Method

mask_circle!(::AbstractMatrix, npix; fill=0)
mask_circle!(::AbstractArray, npix; fill=0)

In-place version of mask_circle

HCIToolbox.mask_circle — Method

mask_circle(::AbstractMatrix, npix; fill=0)
mask_circle(::AbstractArray, npix; fill=0)

Mask the inner-circle of an image with radius npix with value fill. Note that the input type must be compatible with the fill value's type. If the input is a cube it will mask each frame individually.

See Also

mask_circle!

HCIToolbox.normalize_par_angles! — Method

normalize_par_angles!(angles)

In-place version of normalize_par_angles

HCIToolbox.normalize_par_angles — Method

normalize_par_angles(angles)

Ensures parallactic angle list (in degrees) is positive monotonic with no jumps greater than 180°.

Examples

julia> normalize_par_angles([-10, 20, 190])
3-element Array{Int64,1}:
 350
  20
 190

HCIToolbox.shift_frame! — Method

shift_frame!(cube, dx, dy; fill=0)
shift_frame!(cube, dpos; fill=0)

In-place version of shift_frame which modifies cube.

HCIToolbox.shift_frame — Method

shift_frame(frame, dx, dy; fill=0)
shift_frame(frame, dpos; fill=0)

Shifts frame by dx and dy with bilinear interpolation. If necessary, empty indices will be filled with fill.

Examples

julia> shift_frame([0 0 0; 0 1 0; 0 0 0], 1, -1)
3×3 Array{Float64,2}:
 0.0  0.0  1.0
 0.0  0.0  0.0
 0.0  0.0  0.0

julia> shift_frame(ans, (-1, 1), fill=NaN)
3×3 Array{Float64,2}:
 NaN    NaN    NaN
   0.0    1.0  NaN
   0.0    0.0  NaN

HCIToolbox.shift_frame — Method

shift_frame(cube, dx, dy; fill=0)
shift_frame(cube, dpos; fill=0)

Shift each frame of cube by dx and dy, which can be integers or vectors. The change in position can be given as a tuple, which can also be put into a vector to use across the cube. If a frame is shifted outside its axes, the empty indices will be filled with fill.

See Also

shift_frame!

HCIToolbox.get_annulus_segments — Method

get_annulus_segments(data, inner_radius, width, [nsegments]; theta_init=0, scale_factor=1, mode=:index)

Returns indices or values in segments of a centered annulus.

Modes

:mask - returns a positive boolean mask for indexing

:value - returns the data values indexed with the boolean mask

:apply - returns the input data weighted by the boolean mask