API/Reference

For now, here is a dump of all documented functions and types.

Index

HCIToolbox.Metrics.significance

HCIToolbox.Metrics.snr

HCIToolbox.Metrics.snr_approx

HCIToolbox.Metrics.snr_approx!

HCIToolbox.Metrics.snrmap

HCIToolbox.combine

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.expand

HCIToolbox.flatten

HCIToolbox.get_annulus_segments

HCIToolbox.mask_annulus

HCIToolbox.mask_annulus!

HCIToolbox.mask_circle

HCIToolbox.mask_circle!

HCIToolbox.rotate

HCIToolbox.rotate!

API/Reference

HCIToolbox.combine — Method

combine(cube, [angles]; method=median)

Combine all the frames of a cube using method. If angles are provided, will use derotate! before combining.

Note that with angles this is slightly slower than doing combine(derotate!(cube, angles)), if you can overwrite the input.

Examples

julia> X = ones(2, 3, 3)
2×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 2] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 3] =
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X, [0, 90])
1×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0

[:, :, 2] =
 NaN  1.0  NaN

[:, :, 3] =
 1.0  1.0  1.0

HCIToolbox.derotate! — Method

derotate!(cube, angles)

In-place version of derotate

HCIToolbox.derotate — Method

derotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i counter-clockwise by the amount deg2rad(angles[i]).

See Also

derotate!, rotate, rotate!

HCIToolbox.expand — Method

expand(matrix)

Given a matrix of size (n, z), returns a cube of size (n, x, x) where x=√z.

Will throw an error if z is not a perfect square.

Examples

julia> X = ones(3, 4)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> expand(X)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

See Also

flatten

HCIToolbox.flatten — Method

flatten(cube)

Given a cube of size (n, x, y) returns a matrix with size (n, x * y) where each row is a flattened image from the cube.

Examples

julia> X = ones(3, 2, 2)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> flatten(X)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

See Also

expand

HCIToolbox.mask_annulus! — Method

mask_annulus!(::AbstractMatrix, npix_in, npix_out; fill=NaN)

In-place version of mask_annulus

HCIToolbox.mask_annulus — Method

mask_annulus(::AbstractMatrix, npix_in, npix_out; fill=0)

Mask an annular region of an image with inner-radius npix_in, outer-radius npix_out with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_annulus!

HCIToolbox.mask_circle! — Method

mask_circle!(::AbstractMatrix, npix; fill=0)

In-place version of mask_circle

HCIToolbox.mask_circle — Method

mask_circle(::AbstractMatrix, npix; fill=0)

Mask the inner-circle of an image with radius npix with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_circle!

HCIToolbox.rotate! — Method

rotate!(cube, angles)

In-place version of derotate

HCIToolbox.rotate — Method

rotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i clockwise by the amount deg2rad(angles[i]).

See Also

rotate!, derotate, derotate!

HCIToolbox.get_annulus_segments — Method

get_annulus_segments(data, inner_radius, width, [nsegments]; theta_init=0, scale_factor=1, mode=:index)

Returns indices or values in segments of a centered annulus.

Modes

:mask - returns a positive boolean mask for indexing

:value - returns the data values indexed with the boolean mask

:apply - returns the input data weighted by the boolean mask

HCIToolbox.Metrics.significance — Method

significance(snr::AbstractMatrix, fwhm)

Calculates the Gaussian significance from the signal-to-noise ratio (SNR, S/N).

The Gaussian signifiance is calculated from converting the SNR confidence limit from a student t distribution to a Gaussian via

$\text{sig}(\text{SNR}) = \Phi^{-1}\left[\int_0^\text{SNR}{t_\nu(x)dx}\right]$

where the degrees of freedom $\nu$ is given as $2\pi r / \Gamma - 2$ where r is the radial distance of each pixel from the center of the frame.

See Also

snrmap

HCIToolbox.Metrics.snr — Method

snr(data, position, fwhm)

Calculate the signal to noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

Uses the method of Mawet et al. 2014 which includes penalties for small sample statistics. These are encoded by using a student's t-test for calculating the SNR.

Note

SNR is not equivalent to significance, use significance instead

HCIToolbox.Metrics.snr_approx — Method

snr_approx(data, position, fwhm)

Calculates an approximate signal to noise ratio (SNR, S/N) for a test point at position using apertures of diameter fwhm in a residual frame.

Applies the small sample statistics penalty the same as snr. Data is assumed to have been filtered using a 2D top-hat kernel already (automatically done if called via snrmap)

Note

SNR is not equivalent to significance, use significance instead.

HCIToolbox.Metrics.snrmap — Method

snrmap(data, fwhm; snr_func=snr)

Parallel implementation of signal-to-noise ratio (SNR, S/N) applied to each pixel in the input image.

SNR Functions

snr - Exact SNR using student t statistics (small samples penalty)

snr_approx! - Approximate SNR using student t statistics (small samples penalty). Using this function will automatically filter data with a 2D top-hat kernel of diameter fwhm. You can use the non-mutating snr_approx, but the data is already copied after filtering so the mutating version will be more performant.

Tip

This code is automatically multi-threaded, so be sure to set JULIA_NUM_THREADS before loading your runtime to take advantage of it!

HCIToolbox.Metrics.snr_approx! — Method

snr_approx!(data, position, fwhm)

In-place version of snr_approx which modifies data.