API/Reference

For now, here is a dump of all documented functions and types.

Index

HCIToolbox.Mean

HCIToolbox.Median

HCIToolbox.NMF

HCIToolbox.PCA

HCIToolbox.Pairet

Base.reduce

HCIToolbox.combine

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.design

HCIToolbox.expand

HCIToolbox.flatten

HCIToolbox.mask_inner

HCIToolbox.mask_inner!

HCIToolbox.rotate

HCIToolbox.rotate!

API/Reference

HCIToolbox.Mean — Type

Mean

Design using the mean of the cube

Examples

julia> cube = ones(30, 100, 100);

julia> design(Mean, cube)
(A = [1.0 1.0 … 1.0 1.0], w = LinearAlgebra.UniformScaling{Bool}(true), S = [1.0 1.0 … 1.0 1.0])

See Also

Median

HCIToolbox.Median — Type

Median

Design using the median of the cube

Examples

julia> cube = ones(30, 100, 100);

julia> design(Median, cube)
(A = [1.0 1.0 … 1.0 1.0], w = LinearAlgebra.UniformScaling{Bool}(true), S = [1.0 1.0 … 1.0 1.0])

See Also

Mean

HCIToolbox.NMF — Type

NMF

Use non-negative matrix factorization (NMF) to reduce data cube.

Uses NMF.nnmf for decomposition.

Arguments

ncomps::Int - The number of components to keep. Cannot be larger than the number of frames in the input cube (default).

Examples

julia> cube = ones(30, 100, 100);

julia> design(NMF, cube)
(A = [0.18699723925131992 0.0017227311439991433 … 0.25104857267542474 0.007639437531209008; 0.17892712604685682 0.0017227311439991433 … 0.5290547113213493 0.0039969810485154245; … ; 0.1791108604483525 0.3333726515006622 … 0.009151289339269827 0.0039969810485154245; 0.18321575873721374 0.3333726515006622 … 0.009151289339269827 2.90943813485493e-13], w = [4.706954006158167 4.706954006158167 … 4.706954006158167 4.706954006158167; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0], S = [0.9999999998346566 0.9999999998346566 … 0.9999999998346566 0.9999999998346566; 1.0000000001363376 1.0000000001363376 … 1.0000000001363376 1.0000000001363376; … ; 1.0000000001296059 1.0000000001296059 … 1.0000000001296059 1.0000000001296059; 0.9999999999760179 0.9999999999760179 … 0.9999999999760179 0.9999999999760179])

julia> design(NMF, cube, 5)
(A = [0.18257421287124714 0.005547894341040443 … 0.09696771072251913 0.15665986349016803; 0.1825741813281496 0.0055478943410398615 … 0.012760078325461317 0.25756981304356596; … ; 0.18257415421552403 0.33337265150066164 … 0.006436229867279868 0.32699785187779695; 0.18257415421552403 0.33337265150066164 … 0.006436229867279868 0.32699785187779695], w = [5.47722321153941 5.47722321153941 … 5.47722321153941 5.47722321153941; 1.0094317927007388e-7 1.0094317927007388e-7 … 1.0094317927007388e-7 1.0094317927007388e-7; … ; 6.241103130392369e-8 6.241103130392369e-8 … 6.241103130392369e-8 6.241103130392369e-8; 1.7433266010891995e-6 1.7433266010891995e-6 … 1.7433266010891995e-6 1.7433266010891995e-6], S = [0.9999999999004122 0.9999999999004122 … 0.9999999999004122 0.9999999999004122; 1.0000000000166112 1.0000000000166112 … 1.0000000000166112 1.0000000000166112; … ; 1.0000000001164895 1.0000000001164895 … 1.0000000001164895 1.0000000001164895; 1.0000000001164895 1.0000000001164895 … 1.0000000001164895 1.0000000001164895])

HCIToolbox.PCA — Type

PCA

Use principal component analysis (PCA) to reduce data cube.

Uses MultivariateStats.PCA for decomposition. See MultivariateStats.fit(PCA; ...) for keyword arguments

Arguments

ncomps::Int - The number of components to keep. Cannot be larger than the number of frames in the input cube (default).

Examples

julia> cube = ones(30, 100, 100);

julia> design(PCA, cube)
(A = [-0.18257418583505536; -0.18257418583505536; … ; -0.18257418583505447; -0.18257418583505555], w = [-5.477225575051664 -5.477225575051664 … -5.477225575051664 -5.477225575051664], S = [1.0000000000000004 1.0000000000000004 … 1.0000000000000004 1.0000000000000004; 1.0000000000000004 1.0000000000000004 … 1.0000000000000004 1.0000000000000004; … ; 0.9999999999999956 0.9999999999999956 … 0.9999999999999956 0.9999999999999956; 1.0000000000000016 1.0000000000000016 … 1.0000000000000016 1.0000000000000016])

julia> design(PCA, cube, 5)
(A = [-0.18257418583505536; -0.18257418583505536; … ; -0.18257418583505447; -0.18257418583505555], w = [-5.477225575051664 -5.477225575051664 … -5.477225575051664 -5.477225575051664], S = [1.0000000000000004 1.0000000000000004 … 1.0000000000000004 1.0000000000000004; 1.0000000000000004 1.0000000000000004 … 1.0000000000000004 1.0000000000000004; … ; 0.9999999999999956 0.9999999999999956 … 0.9999999999999956 0.9999999999999956; 1.0000000000000016 1.0000000000000016 … 1.0000000000000016 1.0000000000000016])

HCIToolbox.Pairet — Type

Pairet{<:Union{PCA, NMF}}

HCIToolbox.combine — Method

combine(cube, [angles]; method=median)

Combine all the frames of a cube using method. If angles are provided, will use derotate! before combining.

Note that with angles this is slightly slower than doing combine(derotate!(cube, angles)), if you can overwrite the input.

Examples

julia> X = ones(2, 3, 3)
2×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 2] =
 1.0  1.0  1.0
 1.0  1.0  1.0

[:, :, 3] =
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> combine(X, [0, 90])
1×3×3 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0  1.0

[:, :, 2] =
 NaN  1.0  NaN

[:, :, 3] =
 1.0  1.0  1.0

HCIToolbox.derotate! — Method

derotate!(cube, angles)

In-place version of derotate

HCIToolbox.derotate — Method

derotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i counter-clockwise by the amount deg2rad(angles[i]).

See Also

derotate!

rotate

rotate!

HCIToolbox.design — Function

design(::Type{<:HCIAlgorithm}, cube, args...; kwargs...)

Create a design matrix and weights from the given cube. The args and kwargs will vary based on the design algorithm.

Returns

The output of a design matrix will be a named tuple with 3 parameters:

A - The design Matrix

w - The weight vector (the transform of our data cube)

S - The reconstruction of our data cube (usually A * w)

HCIToolbox.expand — Method

expand(matrix)

Given a matrix of size (n, z), returns a cube of size (n, x, x) where x=√z.

Will throw an error if z is not a perfect square.

Examples

julia> X = ones(3, 4)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> expand(X)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

See Also

flatten

HCIToolbox.flatten — Method

flatten(cube)

Given a cube of size (n, x, y) returns a matrix with size (n, x * y) where each row is a flattened image from the cube.

Examples

julia> X = ones(3, 2, 2)
3×2×2 Array{Float64,3}:
[:, :, 1] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

[:, :, 2] =
 1.0  1.0
 1.0  1.0
 1.0  1.0

julia> flatten(X)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

See Also

expand

HCIToolbox.mask_inner! — Method

mask_inner!(::AbstractMatrix, npix; fill=NaN)

In-place version of mask_inner

HCIToolbox.mask_inner — Method

mask_inner(::AbstractMatrix, npix; fill=NaN)

Mask the inner-circle of an image with radius npix with value fill. Note that the input type must be compatible with the fill value's type.

See Also

mask_inner!

HCIToolbox.rotate! — Method

rotate!(cube, angles)

In-place version of derotate

HCIToolbox.rotate — Method

rotate(cube, angles)

Rotates an array using the given angles in degrees.

This will rotate frame i clockwise by the amount deg2rad(angles[i]).

See Also

rotate!

derotate

derotate!

Base.reduce — Method

reduce(::Type{<:HCIAlgorithm}, cube, angles, args...; method=median, kwargs...)

Using a given HCIAlgorithm, will reduce the cube by first finding the approximate reconstruction with design and then derotating and combining (using whichever function specified by method). Any kwargs will be passed to design.