Signal Injection

The following functions are used for injecting fake signal into HCI data.

Index

HCIToolbox.CubeGenerator

HCIToolbox.inject

HCIToolbox.inject!

API/Reference

HCIToolbox.inject — Function

inject(frame, psf, position; A=1, degree=Linear(), fill=0.0)

Injects the matrix psf into frame at the given position. If position is a Tuple, it will be parsed as pixel coordinates (x, y). If position is a Polar (from CoordinateTransformations.jl) it will be parsed as polar coordinates centered on the frame center.

The amplitude of the injected frame can be changed with A, and in the case that psf is dense matrix, and extrapolator will be created with the given degree and fill value.

Coordinate System

Cartesian - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

Polar - Parsed as polar coordinates from the center of the image. theta is a position angle.

Note

Due to the integral nature of array indices, frames or images with even-sized axes will have their center rounded to the nearest integer. This may cause unexpected results for small frames and images.

Examples

julia> inject(zeros(5, 5), ones(1, 1), (2, 1); A=2) # image coordinates (x=2, y=1)
5×5 Array{Float64,2}:
 0.0  2.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> inject(zeros(5, 5), ones(3, 3), Polar(1.5, deg2rad(90))) # polar coords
5×5 Array{Float64,2}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  1.0  1.0  0.0
 0.0  1.0  1.0  1.0  0.0

inject(cube, [angles], psf, position; A=1, degree=Linear(), fill=0.0)

Injects the matrix psf into each frame of cube at the given position. If position is a Tuple, it will be parsed as pixel coordinates (x, y). If position is a Polar (from CoordinateTransformations.jl) it will be parsed as polar coordinates centered on the frame center.

If angles are provided, each position will be rotated by the give parallactic angles in degrees.

The amplitude of the injected frame can be changed with A, and in the case that psf is dense matrix, an extrapolator will be created with the given degree and fill value.

HCIToolbox.inject! — Function

inject!(cube, [angles], psf, position; A=1, degree=Linear(), fill=0.0)

In-place version of inject which modifies cube.

HCIToolbox.CubeGenerator — Type

CubeGenerator(cube, angles, psf; degree=Linear(), fill=0)

Creates a generator which can rapidly create synthetic data using a PSF. The cube and angles are used to define the geometry of the synthetic data cube. If the psf is a dense matrix, an extrapolator will be created with the given degree and fill value.

Examples

julia> using HCIDatasets: BetaPictoris

julia> cube, angles, psf = BetaPictoris[:cube, :pa, :psf];

julia> gen = CubeGenerator(cube, angles, psf); # using empirical PSF

julia> using PSFModels: Gaussian

julia> gen2 = CubeGenerator(cube, angles, Gaussian(4.7)); # using PSFModel

(::CubeGenerator)([T], pos; A=1)
(::CubeGenerator)(base, pos; A=1)

Use the cube generator to generate a synthetic cube. The eltype can be manually specified with T, or otherwise a base can be given. If a base is given, the synthetic cube will be injected, which will add it to the base. base can be a matrix, in which case the generator will automatically flatten the output without extra allocations.

In the case that the generator is build on top of a geometry like AnnulusView, using a matrix base will use the filtered matrix in order to avoid calculating out of bounds indices.

Examples

julia> using HCIDatasets: BetaPictoris

julia> cube, angles, psf = BetaPictoris[:cube, :pa, :psf];

julia> gen = CubeGenerator(cube, angles, psf); # using empirical PSF

julia> s = gen((51, 51)); # cube with PSF at center

julia> size(s) == size(cube)
true

julia> flat_s = gen(zero(flatten(cube)), (51, 51));

julia> flat_s ≈ flatten(s)
true

julia> gen(Polar(10, deg2rad(45)); A=10); # inject using Polar coords