Signal Injection

The following functions are used for injecting fake signal into HCI data.

Index

HCIToolbox.Kernels

HCIToolbox.Kernels.AiryDisk

HCIToolbox.Kernels.Gaussian

HCIToolbox.Kernels.Moffat

HCIToolbox.construct

HCIToolbox.inject

HCIToolbox.inject!

API/Reference

HCIToolbox.inject — Function

inject(frame, ::PSFKernel; A=1, location...)
inject(frame, ::AbstractMatrix; A=1, degree=Linear(), location...)

Injects the given PSF kernel or image into frame with amplitude A. The location can be given in image or polar coordinates, following the coordinate convention below. If no location is given, will assume the center of the frame. For empirical PSFs, degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

Note

Due to the integral nature of array indices, frames or images with even-sized axes will have their center rounded to the nearest integer. This may cause unexpected results for small frames and images.

Examples

julia> inject(zeros(5, 5), ones(1, 1), A=2, x=2, y=1) # image coordinates
5×5 Array{Float64,2}:
 0.0  2.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> inject(zeros(5, 5), ones(3, 3), r=1.5, theta=90) # polar coords
5×5 Array{Float64,2}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  1.0  1.0  0.0
 0.0  1.0  1.0  1.0  0.0

inject(cube, ::PSFKernel, [angles]; A=1, location...)
inject(cube, ::AbstractMatrix, [angles]; A=1, degree=Linear(), location...)

Injects A * img into each frame of cube at the position given by the keyword arguments. If angles are provided, the position in the keyword arguments will correspond to the img position on the first frame of the cube, with each subsequent repositioned img being rotated by -angles in degrees. This is useful for fake companion injection. If no location is given, will assume the center of each frame. For empirical PSFs, degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

HCIToolbox.inject! — Function

inject!(frame, ::PSFKernel; A=1, location...)
inject!(frame, ::AbstractMatrix; A=1, degree=Linear(), location...)

In-place version of inject which modifies frame.

inject!(cube, ::PSFKernel, [angles]; A=1, location...)
inject!(cube, ::AbstractMatrix, [angles]; A=1, location...)

In-place version of inject which modifies cube.

HCIToolbox.construct — Function

construct(kernel, size; A=1, pa=0, location...)
construct(kernel, indices; A=1, pa=0, location...)

Constructs a synthetic PSF using the given kernel function.

The kernel functions are expected to take the distance from the center of the PSF and return the density. Kernels contains common optical PSFs. These kernels are normalized such that the peak amplitude is 1.

If indices are provided they will be used as the input grid. If size is provided the PSF will have the given size. It will have amplitude A and will be located at the given location (explained below) potentially rotated by an additional pa degrees clockwise. If no location is given, will assume the center of the frame.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

construct(::AbstractMatrix, size; A=1, pa=0, degree=Linear(), location...)
construct(::AbstractMatrix, indices; A=1, pa=0, degree=Linear(), location...)

Constructs a PSF at the given location using the given matrix via bilinear interpolation.

If indices are provided they will be used as the input grid. If size is provided the PSF will have the given size. It will have amplitude A and will be located at the given location (explained below) potentially rotated by an additional pa degrees clockwise. If no location is given, will assume the center of the frame. degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

HCIToolbox.Kernels — Module

This module contains some synthetic PSF kernels.

Available kernels

Kernels.Gaussian/Kernels.Normal

Kernels.Moffat

Kernels.AiryDisk

HCIToolbox.Kernels.Gaussian — Type

Kernels.Gaussian(fwhm)
Kernels.Normal(fwhm)

A Gaussian kernel

\[K(d) = \exp\left[-4\ln{2}\left(\frac{d}{\Gamma}\right)^2\right]\]

  ┌────────────────────────────────────────┐
1 │                  .'::                  │ Gaussian(x)
  │                 .: :'.                 │
  │                 :  : :                 │
  │                :'  : ':                │
  │                :   :  :                │
  │               .'   :  ':               │
  │               :    :   :               │
  │              .:    :   '.              │
  │              :     :    :              │
  │             .'     :    ':             │
  │             :      :     :             │
  │            .'      :     ':            │
  │           .'       :      ':           │
  │          .'        :       ':          │
0 │.........''         :         ':........│
  └────────────────────────────────────────┘
  -2fwhm                               2fwhm

HCIToolbox.Kernels.AiryDisk — Type

Kernels.AiryDisk(fwhm)

An Airy disk. Guaranteed to work even at r=0.

\[K(r) = \left[\frac{2J_1\left(\pi r \right)}{\pi r}\right]^2 \quad r \approx \frac{d}{0.97\Gamma}\]

  ┌────────────────────────────────────────┐
1 │                  .'::                  │ AiryDisk(x)
  │                 .' :':                 │
  │                 :  : :                 │
  │                :'  : ':                │
  │                :   :  :                │
  │               .'   :  ':               │
  │               :    :   :               │
  │              .'    :   ':              │
  │              :     :    :              │
  │             .:     :    '.             │
  │             :      :     :             │
  │            .'      :     ':            │
  │            :       :      :.           │
  │           :        :       :           │
0 │..........:'        :        '..........│
  └────────────────────────────────────────┘
  -2fwhm                               2fwhm

HCIToolbox.Kernels.Moffat — Type

Kernels.Moffat(fwhm)

A Moffat kernel

\[K(d) = \left[1 + \left(\frac{d}{\Gamma/2}\right)^2 \right]^{-1}\]

  ┌────────────────────────────────────────┐
1 │                  .::.                  │ Moffat(x)
  │                  : ::                  │
  │                 :' :':                 │
  │                 :  : :.                │
  │                :   :  :                │
  │               .:   :  :.               │
  │               :    :   :               │
  │              .:    :   '.              │
  │              :     :    '.             │
  │             :      :     :.            │
  │            :       :      :.           │
  │          .'        :       ':          │
  │       ..''         :         ':.       │
  │ ....:''            :            ''.... │
0 │''                  :                  '│
  └────────────────────────────────────────┘
  -2fwhm                               2fwhm