Processing

There are a variety of functions defined by HCIToolbox to process HCI data. In particular, routines like derotating cubes, collapsing data cubes, scaling SDI data, etc. are available.

Multi-threading

Many of the methods that work on cubes of data try to multi-thread the operations along the time axis. Make sure you set the environment variable JULIA_NUM_THREADS before staring your runtime to take advantage of this.

Index

HCIToolbox.collapse

HCIToolbox.collapse!

HCIToolbox.construct

HCIToolbox.crop

HCIToolbox.cropview

HCIToolbox.derotate

HCIToolbox.derotate!

HCIToolbox.expand

HCIToolbox.flatten

HCIToolbox.inject

HCIToolbox.inject!

HCIToolbox.invscale

HCIToolbox.invscale_and_collapse

HCIToolbox.scale

HCIToolbox.scale_and_stack

HCIToolbox.shift_frame

HCIToolbox.shift_frame!

API/Reference

HCIToolbox.collapse — Function

collapse(cube; method=median, fill=0, degree=Linear())
collapse(cube, angles; method=:deweight, fill=0, degree=Linear())

Combine all the frames of a cube using method. If angles are provided, will use derotate before combining.

If method is :deweight, the method of Bottom et al. 2017 will be used in which the combined image will be the derotated weighted sum of the frames weighted by the temporal variance. Keyword arguments will be passed to derotate.

References

Bottom et al. 2017 "Noise-weighted Angular Differential Imaging"

Examples

julia> X = ones(2, 3, 3);

julia> collapse(X)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

julia> collapse(X, [0, 90])
3×3 Array{Float64,2}:
 0.5  1.0  0.5
 1.0  1.0  1.0
 0.5  1.0  0.5

julia> collapse(X, [0, 90], fill=NaN)
3×3 Array{Float64,2}:
 NaN    1.0  NaN
   1.0  1.0    1.0
 NaN    1.0  NaN

See Also

collapse!

HCIToolbox.collapse! — Function

collapse!(cube, angles; method=:deweight, fill=0)

An in-place version of the derotating collapse. The only difference is in this version the cube will be derotated in-place.

HCIToolbox.crop — Function

crop(frame, size; center=center(frame), force=false)
crop(cube, size; center=center(cube), force=false)

Crop a frame or cube to size. size can be a tuple or an integer, which will make a square crop. The indices will be relative to center. To avoid uneven (odd) cropping, we may change size. To disable this behavior, set force to true. To avoid allocations, consider cropview.

HCIToolbox.cropview — Function

cropview(cube::AbstractArray{T, 3}, size; center=center(cube), force=false)

Crop a frame to size, returning a view of the frame. size can be a tuple or an integer, which will make a square crop. The indices will be relative to center. To avoid uneven (odd) cropping, we may change size. To disable this behavior, set force to true.

See Also

crop

cropview(frame::AbstractMatrix, size; center=center(frame), force=false)

Crop a frame to size, returning a view of the frame. size can be a tuple or an integer, which will make a square crop. The indices will be relative to center. To avoid uneven (odd) cropping, we may change size. To disable this behavior, set force to true.

See Also

crop

HCIToolbox.derotate — Function

derotate(frame, angle; fill=0, degree=Linear())

Rotates frame counter-clockwise by angle, given in degrees. This is merely a convenient wrapper around ImageTransformations.imrotate.

derotate(cube, angles; fill=0, degree=Linear())

Rotates an array using the given angles in degrees.

This will rotate frame i counter-clockwise. Any values outside the original axes will be replaced with fill. If the given angles are true parallactic angles, the resultant cube will have each frame aligned with top pointing North. degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

Examples

julia> X = zeros(1, 3, 3); X[1, 1, 2] = 1;

julia> X[1, :, :]
3×3 Array{Float64,2}:
 0.0  1.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

julia> derotate(X, [90])[1, :, :]
3×3 Array{Float64,2}:
 0.0  3.22941e-16  0.0
 0.0  0.0          0.999991
 0.0  0.0          0.0

See Also

derotate!

HCIToolbox.derotate! — Function

derotate!(cube, angles; fill=0, degree=Linear())

In-place version of derotate which modifies cube.

HCIToolbox.expand — Function

expand(matrix)

Given a matrix of size (n, z), returns a cube of size (n, x, x) where x=√z.

Will throw an error if z is not a perfect square.

Examples

julia> X = ones(3, 4)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> expand(X)[1, :, :]
2×2 Array{Float64,2}:
 1.0  1.0
 1.0  1.0

See Also

flatten

HCIToolbox.flatten — Function

flatten(cube)

Given a cube of size (n, x, y) returns a matrix with size (n, x * y) where each row is a flattened image from the cube.

Examples

julia> X = ones(3, 2, 2);

julia> flatten(X)
3×4 Array{Float64,2}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

See Also

expand

HCIToolbox.shift_frame — Function

shift_frame(frame, dx, dy; fill=0)
shift_frame(frame, dpos; fill=0)

Shifts frame by dx and dy with bilinear interpolation. If necessary, empty indices will be filled with fill.

Examples

julia> shift_frame([0 0 0; 0 1 0; 0 0 0], 1, -1)
3×3 Array{Float64,2}:
 0.0  0.0  1.0
 0.0  0.0  0.0
 0.0  0.0  0.0

julia> shift_frame(ans, (-1, 1), fill=NaN)
3×3 Array{Float64,2}:
 NaN    NaN    NaN
   0.0    1.0  NaN
   0.0    0.0  NaN

shift_frame(cube, dx, dy; fill=0)
shift_frame(cube, dpos; fill=0)

Shift each frame of cube by dx and dy, which can be integers or vectors. The change in position can be given as a tuple, which can also be put into a vector to use across the cube. If a frame is shifted outside its axes, the empty indices will be filled with fill.

See Also

shift_frame!

HCIToolbox.shift_frame! — Function

shift_frame!(cube, dx, dy; fill=0)
shift_frame!(cube, dpos; fill=0)

In-place version of shift_frame which modifies cube.

Signal Injection

HCIToolbox.inject — Function

inject(frame, ::PSFKernel; A=1, location...)
inject(frame, ::AbstractMatrix; A=1, degree=Linear(), location...)

Injects the given PSF kernel or image into frame with amplitude A. The location can be given in image or polar coordinates, following the coordinate convention below. If no location is given, will assume the center of the frame. For empirical PSFs, degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

Note

Due to the integral nature of array indices, frames or images with even-sized axes will have their center rounded to the nearest integer. This may cause unexpected results for small frames and images.

Examples

julia> inject(zeros(5, 5), ones(1, 1), A=2, x=2, y=1) # image coordinates
5×5 Array{Float64,2}:
 0.0  2.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0

julia> inject(zeros(5, 5), ones(3, 3), r=1.5, theta=90) # polar coords
5×5 Array{Float64,2}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  1.0  1.0  0.0
 0.0  1.0  1.0  1.0  0.0

inject(cube, ::PSFKernel, [angles]; A=1, location...)
inject(cube, ::AbstractMatrix, [angles]; A=1, degree=Linear(), location...)

Injects A * img into each frame of cube at the position given by the keyword arguments. If angles are provided, the position in the keyword arguments will correspond to the img position on the first frame of the cube, with each subsequent repositioned img being rotated by -angles in degrees. This is useful for fake companion injection. If no location is given, will assume the center of each frame. For empirical PSFs, degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

HCIToolbox.inject! — Function

inject!(frame, ::PSFKernel; A=1, location...)
inject!(frame, ::AbstractMatrix; A=1, degree=Linear(), location...)

In-place version of inject which modifies frame.

inject!(cube, ::PSFKernel, [angles]; A=1, location...)
inject!(cube, ::AbstractMatrix, [angles]; A=1, location...)

In-place version of inject which modifies cube.

HCIToolbox.construct — Function

construct(kernel, size; A=1, pa=0, location...)
construct(kernel, indices; A=1, pa=0, location...)

Constructs a synthetic PSF using the given kernel function.

The kernel functions are expected to take the distance from the center of the PSF and return the density. Kernels contains common optical PSFs. These kernels are normalized such that the peak amplitude is 1.

If indices are provided they will be used as the input grid. If size is provided the PSF will have the given size. It will have amplitude A and will be located at the given location (explained below) potentially rotated by an additional pa degrees clockwise. If no location is given, will assume the center of the frame.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

construct(::AbstractMatrix, size; A=1, pa=0, degree=Linear(), location...)
construct(::AbstractMatrix, indices; A=1, pa=0, degree=Linear(), location...)

Constructs a PSF at the given location using the given matrix via bilinear interpolation.

If indices are provided they will be used as the input grid. If size is provided the PSF will have the given size. It will have amplitude A and will be located at the given location (explained below) potentially rotated by an additional pa degrees clockwise. If no location is given, will assume the center of the frame. degree is the corresponding Interpolations.Degree for the B-Spline used to subsample the pixel values.

Coordinate System

x, y - Parsed as distance from the bottom-left corner of the image. Pixel convention is that (1, 1) is the center of the bottom-left pixel increasing right and up.

r, theta or r, θ - Parsed as polar coordinates from the center of the image. theta is a position angle.

SDI

HCIToolbox.invscale — Function

invscale(cube::AbstractArray{T,3}, scales[, size])

Linearly contract each frame in cube by the corresponding scale in scales. Uses bilinear interpolaiton internally. The output size can be specified or else we will choose the smallest size that contains the largest stretch.

invscale(frame::AbstractMatrix, scale[, size])

Linearly contract frame with the ratio scale. Uses bilinear interpolation internally. If size is specified, will return a crop view.

HCIToolbox.invscale_and_collapse — Function

invscale_and_collapse(stack_cube::AbstractArray{T,3}, scales, size; kwargs...)

Given an SDI tensor that has been stacked into a cube, invscales each spectral slice and combines with collapse. The output will be cropped to size.

HCIToolbox.scale — Function

scale(cube::AbstractArray{T,3}, scales[, size])

Linearly stretch each frame in cube by the corresponding scale in scales. Uses bilinear interpolaiton internally. The output size can be specified or else we will choose the smallest size that contains the largest stretch.

scale(frame::AbstractMatrix, scale[, size])

Linearly stretch frame with the ratio scale, padding symmetrically with zeros to match out_size. Uses bilinear interpolation internally. If size is specified, will return a padded view.

HCIToolbox.scale_and_stack — Function

scale_and_stack(spcube::AbstractArray{T,4}, scales)

Given a 4-D spectral ADI (SDI) tensor, scales each, temporal slice according to scales and then concatenates into a cube with nλ * nf frames.